Add mods parameter to character callback.
This is needed to correctly and _reliably_ distinguish character events
from mod+key combinations, due to the change in #40.
I am sick of improvements behind held back due to "backwards
compatibility". Why is it acceptable to do so much more work to jump
through hoops to use sub-optimal API and still suffer with sub-optimal
results, rather than doing so much less work to improve the API and
code that uses it and have correct results? -.-