Bug fixes to filters.c: All windows now show same contents regardless of position.
The auxiliary windows now reflect what is happening in the primary window.
Unfortunately, there is still a bug in the primary window code that affects
the orthographic projection.  Basically, the object that is drawn is resized
along with the contents of the window instead of remaining a constant size
while the window is resized around it.