Changed the location of one vertex so the triangle is more triangular (maybe?)
I haven't checked the math on this, but in playing with blending modes I noticed
that one of the rotated vertices appeared to be closer to the center of rotation
than the others.  This adjustment makes it so they are all the same distance from
the center of rotation.  If the center of rotation is conincident with the
triangle's center, then this will mean the triangle is equilateral, but I
haven't verified this, it just looked right from the rotations.