Added support for joystick button and axis callbacks
The major change is the addition of the inputValueCallback function.  This callback function is registered with the HID Manager using the HIDDeviceRef when a joystick is connected.  Whenever an axis, button, or hat switch value is changed for that HID device, the inputValueCallback function is called: the appropriate value will be updated (which alleviates the need for polling, hence the removal of the polling functions) and, if a user callback has been registered, event notification will be provided.  (Note that in the case of hat switches, which are treated as quad sets of buttons, one for each direction, user notification is provided only if the value for that direction has actually changed).