Handle X11 Selection* events despite NULL window
`processEvent` in `x11_window.c` currently discards events that can not be mapped to a current GLFW window. However, this breaks clipboard functionality by failing to respond to SelectionRequest and SelectionClear events.

This commit moves processing of these important clipboard events to before the NULL window test so that they are always considered.

Fixes #961 (tentatively; there is probably an underlying structural change to be made that fixes this properly).