Add CMake object library target as well
Some projects (e.g. raylib, nanogui) ship a static library that has GLFW
embedded. Unfortunately, it's not possible to do so portably with the
current CMake setup.

The proper way seems to be to define a CMake OBJECT library target and use
that as source for the glfw target.

User code that is interested in the glfw target is unaffected and can
keep using it normally.
User code that wants the objects to link into a static library just
needs to supply $<TARGET_OBJECTS:glfw_objlib> as part of the source.