Monitor: simplify gamma range check
Assertions and tests in glfwSetGamma verify that user-supplied gamma is
a positive finite value, but tests for error reporting are redundant
given the preceding asserts (however asserts allow gamma == 0, which is
probably not intended).

Remove the asserts and simplify the range check. Alternatively, it may
be spelled as 'isfinite(gamma) && gamma > 0'.