input: Fix issues with XInput on Windows
glfwInit() will call _glfwPlatformInit(), which on Win32 will call
_glfwInitJoysticksWin32(), which will try to create an XInput gamepad
using the SDL device ID for XInput. However, at this point in time, the
default mappings for controllers hasn't been initialized, so we end up
with an XInput gamepad with no mappings.

Fix this by loading the mappings before calling glfwInit().