Fix #1112: simultaneous key presses being dropped
This improves the mechanism to check for duplicate key press events, by
checking if the event with the same time code actually corresponds to
the same key.

This will correctly process any key known to GLFW, handling simultanous
presses of different keys, while still filtering out duplicate presses
of the same key.

Note that this still does not properly handle keys that are
GLFW_KEY_UNKNOWN, as I am not aware of a good way to check for previous
events on those keys.