Query and keep track of Xkb group index
Ubuntu 18.04: for users with multiple keyboard layouts configured,
glfwGetKeyName works fine only with the primary layout. Switching
layouts results in changing the group index. This commit querries the
current group index when initializing keyboard input and keeps track of
any change to it.

As a result the scancode -> keyname mapping may change while the program
is running (needs to be documented).

This needs to be tested on other linux distros.