input: Cache parsed mappings.h
Parsing this file was taking the most time out of glfwInit(), or about
14ms on my PinePhone before the start of this series.  In the case where
the string is already in cache, it now takes only 0.95ms.

It would be possible to make that even faster, and additionally to keep
that page shared with other GLFW processes, by using mmap() instead, but
for now I opted to not change the memory representation.