Enable joystick support on FreeBSD using the Linux driver.
This is arguably incomplete; it supports joysticks via webcamd (which uses Linux
drivers running in userspace), but not via uhidd (which is FreeBSD's own thing).

Nevertheless a strict improvement.

Tested using a gamepad I have that is supported by the "xpad" Linux driver.

Note: adds build dependency on devel/evdev-proto and devel/libinotify.
To actually use it, need to install multimedia/webcamd at runtime (if not
installed or not running, no joysticks will be detected).

Fixes #1892.