X11: Fix segfault when _glfw.x11.display is higher than 1024
FD_SETSIZE is hardcoded to 1024 in glibc. See the bugs section of the select call for more info.
In some cases, it is possible for _glfw.x11.display to be larger than FD_SETSIZE, causing a segfault or buffer overflow.
Replacing select() with poll() seems to be the recommended solution:
- https://www.man7.org/linux/man-pages/man2/select.2.html#BUGS
- https://access.redhat.com/solutions/488623
- https://stackoverflow.com/questions/7976388/increasing-limit-of-fd-setsize-and-select/7977082#7977082